Move GhidraTypeBuilder.kt to analysis.objectivec
